--- pkg/engine.go
@@ -1,7 +1,0 @@
-package pkg
-
-import "context"
-
-type LoadTester interface {
-	Start(ctx context.Context) error
-}
