+++ pkg/engine/engine.go
@@ -1,0 +1,7 @@
+package engine
+
+import "context"
+
+type WorkloadEvaluator interface {
+	StressAndBenchmark(ctx context.Context) error
+}
